Винесення функціоналу в utils.py
через помилку кругового імпорту, не вдалося винести в utils.py функціонал збереження url та функції валідації. image_handler.py переніс до utils.py
